refactor(front): moves globals, renames styles directory, and adjusts imports
